photosec_demo/requirements.txt
asgiref==3.2.7
astroid==2.3.3
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0.5
django-import-export==2.0.2
django-rest-swagger==2.2.0
djangorestframework==3.11.0
drf-yasg==1.17.1
et-xmlfile==1.0.1
idna==2.9
inflection==0.4.0
isort==4.3.21
itypes==1.1.0
jdcal==1.4.1
Jinja2==2.11.2
jsonschema==3.2.0
lazy-object-proxy==1.4.3
MarkupPy==1.14
MarkupSafe==1.1.1
mccabe==0.6.1
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.0.3
packaging==20.3
pylint==2.4.4
pyparsing==2.4.7
pyrsistent==0.16.0
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.1
swagger-spec-validator==2.5.0
tablib==1.1.0
uritemplate==3.0.1
urllib3==1.25.8
virtualenv==16.7.9
wrapt==1.11.2
xlrd==1.2.0
xlwt==1.3.0
